Added Revision class and refactored the Repository class
